Fixed issue with menus not showing, simplified ability system.
The issue with menus not displaying was due to a bad line of code in
MapScreen resize handler.

Instead of stage.setViewPort(...), the proper way to handle it is via
stage.getViewPort().update(width, height).

Changed so that stage now initializes with a FitViewport, in order to
keep the menu aspect ratio on resize.

Simplified the ability system, removed the concept of MP cost. Made two
placeholder abilities, an attack and a heal.